requirements.txt
google-api-python-client>=2.120.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
python-dotenv>=1.0.1
requests>=2.31.0
